updated mvvm structure
