+++ Quicker/Models/CsvFile.cs
@@ -1,0 +1,38 @@
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Text;
+using System.Text.RegularExpressions;
+using System.Threading.Tasks;
+using System.Windows.Shapes;
+
+namespace Quicker.Models
+{
+    internal class CsvFile
+    {
+        private string CsvPath;
+        public CsvFile(string path)
+        {
+            this.CsvPath = path;
+        }
+
+        public Dictionary<string, Match> ReadCsv()
+        {
+            var Matches = new Dictionary<string, Match>();
+            using (var reader = new StreamReader(CsvPath))
+            {
+                while (!reader.EndOfStream)
+                {
+                    var line = reader.ReadLine();
+                    var values = line.Split(',');
+                    var keyword = values[0].Trim();
+                    var snippet = values[1].Trim();
+                    Match match = new Match(keyword, snippet);
+                    Matches.TryAdd(values[0].Trim(), match);
+                }
+            }
+            return Matches;
+        }
+    }
+}
